+++ Changelog.md
@@ -4,4 +4,5 @@
 ## v0.2.2
 
 * Started a Changelog, let's see how long I keep it going.
+* Added `to_usize` method.
 * Added read only and write only variants.